Complete re-write of vimrc
Trying to locate the source of the lag when in insert mode.

After much testing and research it seems to be vims built in Ruby syntax
checking that uses regex. Somewhere after vim v7 a new regex engine was
added that does not like the Ruby syntax regex.

Switching to `set regexpengine=1` seems to be the only thing that stops
the lag.

As this work also cleaned up my vimrc I will keep it and monitor the
performance.