Add Rspec defaults
I loves me the Rspec documentation format, but not everyone does. By
having an `.rpsec` file in my HOME I can set defaults just for me for
all projects.

I read you could base the format on the number of specs, but I have
never found a way to do that and might have imagined it, here I am
basing the format on the numbers of arguments, which generally:

- is none when you call `rspec` or run everything
- and some when you call `rspec this/spec/file_spec.rb`

So this kind of does what I was looking for.

We'll see how it feels.